+++ app/imports/components/appLayout.jsx
@@ -1,11 +1,14 @@
 import React from 'react'
 
+import Header from '/imports/components/header.jsx';
+import Footer from '/imports/components/footer.jsx';
+
 export const AppLayout = ({header, content, footer}) => (
   <div id="app-container">
-    {header}
+    <Header />
     <div className="container">
       {content}
     </div>
-    {footer}
+    <Footer />
   </div>
 );
